deepfake-video-detection-main/backend/requirements.txt
asgiref==3.4.1
click==8.0.3
cmake==3.22.1
colorama==0.4.4
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.22.0
opencv-python==4.5.5.62
Pillow==9.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
tzdata==2021.5
